+++ psutil/_pslinux.py
@@ -783,7 +783,10 @@ class Process(object):
         fname = "/proc/%s/cmdline" % self.pid
         kw = dict(encoding=DEFAULT_ENCODING) if PY3 else dict()
         with open(fname, "rt", **kw) as f:
-            return [x for x in f.read()[:-1].split('\x00')]
+            data = f.read()
+        if data.endswith('\x00'):
+            data = data[:-1]
+        return [x for x in data.split('\x00')]
 
     @wrap_exceptions
     def terminal(self):