+++ HISTORY.rst
@@ -9,7 +9,8 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
 - #634: [Linux] Proces.cmdline() does not include empty string arguments.
 - #635: [UNIX] crash on module import if 'enum' package is installed on python
   < 3.4.
-
+- #636: [Linux] *connections functions may swallow errors and return an
+  incomplete list of connnections.
 
 3.0.0 - 2015-06-13
 ==================